backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
motor==3.6.0
python-dotenv==1.0.1
google-generativeai==0.8.3
